27/11/2021 => Mitigando Bug de Consulta de Usuários

+++ ControleAmbientes/ControleAmbientes/View/Users.xaml.cs
@@ -13,6 +13,8 @@ namespace ControleAmbientes.View
         public Users()
         {
             InitializeComponent();
+
+            dataGridUsers.DataContext = crud.FillDataGrid();
         }
 
         private void Label_MouseUp(object sender, System.Windows.Input.MouseButtonEventArgs e)
@@ -22,12 +24,33 @@ namespace ControleAmbientes.View
 
         private void btDelete_Click(object sender, RoutedEventArgs e)
         {
+            if (txtEmail.Text.Equals("") || txtName.Text.Equals("") || txtLogin.Text.Equals(""))
+            {
+                MessageBox.Show("Todos os campos devem ser preenchidos.", "! ! ! Atenção ! ! !", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            if (!txtPassword.Password.Equals(txtPasswordConfirma.Password) || txtPassword.Password.Equals("") || txtPasswordConfirma.Password.Equals(""))
+            {
+                MessageBox.Show("Falha no cadastro da Senha.", "! ! ! Atenção ! ! !", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
+            string[] splitEmail = txtEmail.Text.Split('@');
+
+            if (splitEmail.Length != 2)
+            {
+                MessageBox.Show("Email no formato inválido.", "! ! ! Atenção ! ! !", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
             crud.DeleteUser(txtEmail.Text);
+
+            dataGridUsers.DataContext = crud.FillDataGrid();
         }
 
         private void btSalve_Click(object sender, RoutedEventArgs e)
         {
-            
 
             if (txtEmail.Text.Equals("") || txtName.Text.Equals("") || txtLogin.Text.Equals(""))
             {
@@ -50,6 +73,8 @@ namespace ControleAmbientes.View
             }
 
             crud.SaveUser(txtEmail.Text, txtName.Text, txtLogin.Text, txtPasswordConfirma.Password.ToString());
+
+            dataGridUsers.DataContext = crud.FillDataGrid();
 
         }
     }
